dev: commit all

+++ backend/app/models/Role.model.ts
@@ -69,28 +69,21 @@ class Role {
   }
 
   static async hasRole (user: User, roles: Roles[]): Promise<boolean> {
-    const result = await Database.query('SELECT * FROM user_roles WHERE user_id = $1 AND role_id = ANY($2)', [user.id, roles])
-    return result.rows.length > 0
+    const result = await Database.query('SELECT roles.name FROM user_roles INNER JOIN roles ON user_roles.role_id = roles.id WHERE user_roles.user_id = $1', [user.id])
+    return roles.some(role => result.rows.some(row => row.name === role))
   }
 
   static async getRoles (user: User): Promise<Role[]> {
-    const result = await Database.query('SELECT role_id FROM user_roles WHERE user_id = $1', [user.id])
-    return result.rows.map(row => row.role_id)
+    const result = await Database.query('SELECT roles.name FROM user_roles INNER JOIN roles ON user_roles.role_id = roles.id WHERE user_roles.user_id = $1', [user.id])
+    return result.rows.map(row => row.name)
   }
 
-  static async addRole (user: User, role: Roles | number): Promise<void> {
-    if (typeof role === 'number') {
-      await Database.query('INSERT INTO user_roles (user_id, role_id) VALUES ($1, $2)', [user.id, role])
-    } else {
-      const roleInstance = await Role.find(role)
-      if (roleInstance) {
-        await Database.query('INSERT INTO user_roles (user_id, role_id) VALUES ($1, $2)', [user.id, roleInstance.id])
-      }
-    }
+  static async addRole (user: User, role: Roles): Promise<void> {
+    await Database.query('INSERT INTO user_roles (user_id, role_id) VALUES ($1, (SELECT id FROM roles WHERE name = $2))', [user.id, role])
   }
 
   static async removeRole (user: User, role: Roles): Promise<void> {
-    await Database.query('DELETE FROM user_roles WHERE user_id = $1 AND role_id = $2', [user.id, role])
+    await Database.query('DELETE FROM user_roles WHERE user_id = $1 AND role_id = (SELECT id FROM roles WHERE name = $2)', [user.id, role])
   }
 }
